C++: Remove the primitive field flow in FlowVar
Now that we have proper flow through fields in the interprocedural data
flow library, it's time to remove the imprecise local flow through
fields in `FlowVar.qll`.

This removes all flow through fields in `DataFlow::localFlow` and other
local-only predicates. The gain is that we no longer get the false
positives that came from ignoring the qualifiers of field accesses.
Performance of `getAReachedBlockVarSBB` is also much improved.